+++ src/main/java/com/motorny/ss/chemistryservice/service/impl/UserServiceImpl.java
@@ -1,12 +1,14 @@
 package com.motorny.ss.chemistryservice.service.impl;
 
 import com.motorny.ss.chemistryservice.dto.UserDto;
+import com.motorny.ss.chemistryservice.exceptions.ResourceNotFoundException;
 import com.motorny.ss.chemistryservice.mapper.UserMapper;
 import com.motorny.ss.chemistryservice.model.Review;
 import com.motorny.ss.chemistryservice.model.User;
 import com.motorny.ss.chemistryservice.repository.UserRepository;
 import com.motorny.ss.chemistryservice.service.UserService;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,6 +17,7 @@ import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
+@Slf4j
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
@@ -32,7 +35,12 @@ public class UserServiceImpl implements UserService {
     @Override
     public UserDto getUser(long id) {
         Optional<User> byId = userRepository.findById(id);
-        User user = byId.orElse(null);
+
+        if (!byId.isPresent()) {
+            throw new ResourceNotFoundException("User not found with id " + id);
+        }
+
+        User user = byId.get();
 
         return userMapper.toUserDto(user);
     }
@@ -46,8 +54,16 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(long id) {
-        userRepository.deleteById(id);
+    public String deleteUser(long id) {
+        Optional<User> byId = userRepository.findById(id);
+
+        if (byId.isPresent()) {
+            userRepository.deleteById(id);
+            return "User with id: " + id + " was successfully remover";
+        } else {
+            log.error("User not found with id {}", id);
+            throw new ResourceNotFoundException("User not found with id " + id);
+        }
     }
 
     @Override
